chore: rename provider.domain
