+++ frontend/dashboard/src/app/[community]/dashboard/layout.tsx
@@ -1,7 +1,7 @@
 import { GlobalLayout } from '~/providers'
 import { getCommunityInfo, getLocaleData } from '~/providers/domain'
 import LocaleStoreProvider from '~/stores/locale.domain/provider'
-import { CommunityInfoProvider } from '~/stores/provider.domain'
+import { CommunityInfoProvider } from '~/stores/provider'
 import Client from './Client'
 
 export default async ({ children, params }) => {
