+++ frontend/dashboard/src/app/apply/community/layout.tsx
@@ -1,7 +1,7 @@
 import { GlobalLayout } from '~/providers'
 import { getCommunityInfo, getLocaleData } from '~/providers/domain'
 import LocaleStoreProvider from '~/stores/locale.domain/provider'
-import { CommunityInfoProvider } from '~/stores/provider.domain'
+import { CommunityInfoProvider } from '~/stores/provider'
 
 export default async ({ children, params }) => {
   const params$ = await params
